+++ packages/react-intersection-observer/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Fixed
+
+- Fix setting state in `IntersectionObserver` render. [#2059](https://github.com/Shopify/quilt/pull/2059)
 
 ## 3.1.6 - 2021-09-24